feat: unify the grid layout of wallets and contacts (#1217)
* feat: use a context menu on contacts grid

* refactor: extract the grid layout from wallets and contacts to \`WalletGrid\`

* refactor: move right-click dropdown to grid
