+++ src/renderer/pages/Wallet/WalletAll.vue
@@ -66,7 +66,7 @@
             <span
               v-if="isLedgerLoading"
               v-tooltip="{
-                content: $t('PAGES.WALLET_ALL.LOADING_LEDGER'),
+                content: $t('WALLET_GRID.LOADING_LEDGER'),
                 placement: 'right'
               }"
               class="inline-flex items-center self-stretch ml-3 pr-2"
@@ -94,106 +94,14 @@
           </div>
         </div>
 
-        <div
+        <WalletGrid
           v-if="hasWalletGridLayout && !isLoading"
-          class="WalletAll__grid mt-10"
-        >
-          <button
-            v-if="isLedgerLoading"
-            :disabled="true"
-            class="WalletAll__grid__wallet"
-          >
-            <Loader />
-            <div class="text-center mt-4">
-              {{ $t('PAGES.WALLET_ALL.LOADING_LEDGER') }}
-            </div>
-          </button>
-
-          <button
-            v-for="wallet in selectableWallets"
-            :key="wallet.id"
-            class="WalletAll__grid__wallet group"
-            @click="showWallet(wallet.id)"
-            @contextmenu.prevent="toggleDropdown(`dropdown-${wallet.id}`)"
-          >
-            <div class="WalletAll__grid__wallet__wrapper">
-              <div class="WalletAll__grid__wallet__wrapper__mask">
-                <div class="flex items-center justify-between">
-                  <div class="flex items-center">
-                    <WalletIdenticon
-                      :value="wallet.address"
-                      :size="60"
-                      class="identicon"
-                    />
-
-                    <div
-                      class="flex flex-col justify-center overflow-hidden pl-4"
-                    >
-                      <div class="flex items-center">
-                        <span
-                          v-tooltip="{
-                            content: !wallet.name && wallet_name(wallet.address) ? $t('COMMON.NETWORK_NAME') : '',
-                            placement: 'right'
-                          }"
-                          class="WalletAll__grid__wallet__name font-semibold text-base truncate block pr-1"
-                        >
-                          {{ wallet.name || wallet_name(wallet.address) || wallet_truncate(wallet.address) }}
-                        </span>
-                        <span
-                          v-if="wallet.isLedger"
-                          class="ledger-badge"
-                        >
-                          {{ $t('COMMON.LEDGER') }}
-                        </span>
-                      </div>
-                      <span
-                        class="font-bold mt-2 text-lg text-theme-page-text text-left whitespace-no-wrap"
-                      >
-                        {{ formatter_networkCurrency(wallet.balance, 2) }}
-                      </span>
-                    </div>
-                  </div>
-
-                  <MenuDropdown
-                    :ref="`dropdown-${wallet.id}`"
-                    :items="getContextMenuOptions(wallet)"
-                    :is-highlighting="false"
-                    :position="['-100%', '-20%']"
-                    :container-classes="'hidden group-hover:block'"
-                    @select="onSelectDropdown(wallet, $event)"
-                  >
-                    <span
-                      slot="handler"
-                      class="WalletAll__grid__wallet__select p-2 text-theme-page-text-light hover:text-theme-page-text opacity-75"
-                    >
-                      <SvgIcon
-                        name="more"
-                        view-box="0 0 5 15"
-                        class="text-inherit"
-                      />
-                    </span>
-
-                    <template
-                      slot="item"
-                      slot-scope="itemScope"
-                    >
-                      <div class="flex items-center hidden">
-                        <SvgIcon
-                          :name="itemScope.item.icon"
-                          view-box="0 0 16 16"
-                          class="text-inherit flex-none mr-2"
-                        />
-                        <span class="font-semibold">
-                          {{ itemScope.item.value }}
-                        </span>
-                      </div>
-                    </template>
-                  </MenuDropdown>
-                </div>
-              </div>
-            </div>
-          </button>
-        </div>
+          :is-ledger-loading="isLedgerLoading"
+          :wallets="selectableWallets"
+          @show="showWallet"
+          @rename="openRenameModal"
+          @remove="openRemovalConfirmation"
+        />
 
         <div
           v-else-if="!hasWalletGridLayout && !isLoading"
@@ -238,9 +146,8 @@ import Loader from '@/components/utils/Loader'
 import { ProfileAvatar } from '@/components/Profile'
 import SvgIcon from '@/components/SvgIcon'
 import { WalletButtonCreate, WalletButtonExport, WalletButtonImport, WalletButtonLedgerSettings } from '@/components/Wallet/WalletButtons'
-import { WalletIdenticon, WalletRemovalConfirmation, WalletRenameModal } from '@/components/Wallet'
+import { WalletGrid, WalletRemovalConfirmation, WalletRenameModal } from '@/components/Wallet'
 import WalletTable from '@/components/Wallet/WalletTable'
-import { MenuDropdown } from '@/components/Menu'
 
 export default {
   name: 'WalletAll',
@@ -254,11 +161,10 @@ export default {
     WalletButtonExport,
     WalletButtonImport,
     WalletButtonLedgerSettings,
-    WalletIdenticon,
+    WalletGrid,
     WalletRemovalConfirmation,
     WalletRenameModal,
-    WalletTable,
-    MenuDropdown
+    WalletTable
   },
 
   data: () => ({
@@ -312,7 +218,7 @@ export default {
     },
 
     isLedgerLoading () {
-      return this.$store.getters['ledger/isLoading'] && !this.$store.getters['ledger/wallets'].length
+      return !!(this.$store.getters['ledger/isLoading'] && !this.$store.getters['ledger/wallets'].length)
     },
 
     isLedgerConnected () {
@@ -429,10 +335,6 @@ export default {
       this.walletLayout = this.walletLayout === 'grid' ? 'tabular' : 'grid'
     },
 
-    toggleDropdown (dropdownId) {
-      this.$refs[dropdownId][0].toggle()
-    },
-
     onRemoveWallet (wallet) {
       this.openRemovalConfirmation(wallet)
     },
@@ -452,32 +354,6 @@ export default {
 
     showWallet (walletId) {
       this.$router.push({ name: 'wallet-show', params: { address: walletId } })
-    },
-
-    getContextMenuOptions (wallet) {
-      const options = {
-        rename: {
-          value: this.$t('WALLET_TABLE.RENAME'),
-          icon: 'edit'
-        }
-      }
-
-      if (!wallet.isLedger) {
-        options['delete'] = {
-          value: this.$t('WALLET_TABLE.DELETE'),
-          icon: 'delete-wallet'
-        }
-      }
-
-      return options
-    },
-
-    onSelectDropdown (wallet, item) {
-      if (item === 'delete') {
-        this.openRemovalConfirmation(wallet)
-      } else if (item === 'rename') {
-        this.openRenameModal(wallet)
-      }
     }
   }
 }
@@ -511,62 +387,5 @@ export default {
 }
 .WalletAll__header {
   @apply .flex .items-center .justify-between .h-8;
-}
-.WalletAll__grid {
-  display: grid;
-  grid-template-rows: 1fr;
-}
-.WalletAll__grid__wallet {
-  @apply py-3 relative cursor-pointer bg-theme-feature;
-  transition-property: transform, border, box-shadow;
-  transition-duration: .2s;
-  transition-timing-function: ease;
-}
-.WalletAll__grid__wallet:hover {
-  @apply rounded-lg z-10;
-  transform: scale(1.02);
-  box-shadow: var(--theme-wallet-grid-shadow);
-}
-.WalletAll__grid__wallet:not(:hover)::after {
-  @apply block absolute pin-x pin-b mx-auto border-b border-theme-wallet-overview-border;
-  content: " ";
-  width: 95%;
-}
-.WalletAll__grid__wallet__wrapper {
-  @apply px-5 py-2 border-l border-theme-wallet-overview-border;
-}
-.WalletAll__grid__wallet:hover .WalletAll__grid__wallet__wrapper {
-  @apply border-transparent
-}
-.WalletAll__grid__wallet:hover .identicon {
-  opacity: 1;
-}
-.WalletAll__grid__wallet__name {
-  color: #037cff;
-}
-.WalletAll__grid__wallet .identicon {
-  opacity: 0.5;
-  transition: 0.5s;
-}
-@screen max-md {
-  .WalletAll__grid__wallet__wrapper {
-    @apply border-transparent
-  }
-}
-@screen minmax-lg {
-  .WalletAll__grid {
-    grid-template-columns: 1fr 1fr;
-  }
-  .WalletAll__grid__wallet:nth-child(2n+1) > .WalletAll__grid__wallet__wrapper {
-    @apply border-transparent
-  }
-}
-@screen xl {
-  .WalletAll__grid {
-    grid-template-columns: 1fr 1fr 1fr;
-  }
-  .WalletAll__grid__wallet:nth-child(3n+1) > .WalletAll__grid__wallet__wrapper {
-    @apply border-transparent
-  }
 }
 </style>
